fix(web): move composio cache invalidation out of route module
Keep Next.js route exports valid by extracting Composio connection caches into a dedicated helper module and updating callback/tests to consume it, which unblocks web production builds.

Made-with: Cursor
